requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
caldav==0.8.0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via
    #   django-encrypted-model-fields
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
django-allauth==0.44.0
    # via -r requirements.in
django-encrypted-model-fields==0.5.8
    # via -r requirements.in
django-friendly-tag-loader==1.3.1
    # via -r requirements.in
django-markdown-app==0.9.6
    # via -r requirements.in
django-pipeline==2.0.6
    # via -r requirements.in
django-simple-csp==0.2.dev1
    # via -r requirements.in
django-token-bucket==0.2.3
    # via -r requirements.in
django-widget-tweaks==1.4.8
    # via -r requirements.in
django==2.2.23
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-encrypted-model-fields
    #   django-markdown-app
    #   django-simple-csp
    #   django-token-bucket
icalendar==4.0.7
    # via -r requirements.in
idna==2.10
    # via requests
libsass==0.21.0
    # via libsasscompiler
libsasscompiler==0.1.8
    # via -r requirements.in
lxml==4.6.3
    # via caldav
markdown==3.3.4
    # via django-markdown-app
oauthlib==3.1.0
    # via requests-oauthlib
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.1.0
    # via django-allauth
python-dateutil==2.8.1
    # via
    #   icalendar
    #   vobject
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
    #   icalendar
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   caldav
    #   django-allauth
    #   requests-oauthlib
six==1.16.0
    # via
    #   caldav
    #   django-encrypted-model-fields
    #   libsass
    #   python-dateutil
sqlparse==0.4.1
    # via django
urllib3==1.26.5
    # via requests
uwsgi==2.0.19.1
    # via -r requirements.in
vobject==0.9.6.1
    # via caldav
